removed electron from the store name and related classes; updated window-state-store.test to use snapshot
