--- src/electron/flux/types/electron-window-state-store-data.ts
@@ -1,8 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-export type ElectronRoutes = 'deviceConnectView' | 'resultsView';
-
-export interface ElectronWindowStateStoreData {
-    routeId: ElectronRoutes;
-}
