+++ src/electron/flux/types/window-state-store-data.ts
@@ -1,7 +1,8 @@
-import { ViewRoutes } from '../types/window-state-store-data';
-
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-export interface RoutePayload {
+
+export type ViewRoutes = 'deviceConnectView' | 'resultsView';
+
+export interface WindowStateStoreData {
     routeId: ViewRoutes;
 }
